+++ Test/other_variants/Lab_7/var_7/dtoa.h
