--- Test/other_variants/Lab_7/var_7/lab07.cpp
@@ -1,27 +1,0 @@
-﻿#include "Header.h"
-#include <stdio.h>
-// Написать функцию для преобразования переменной типа double в строку символов с заданной точностью
-// представления(dtoa(double num, int npos, int nfr)) и тестирующую программу к ней, где npos— -
-// общее число знаков(включая запятую), а nfr— - число знаков после запятой.
-
-int main() {
-	int npost = 0;
-	int nfr = 0;
-	double num;
-	scanf("%lf", &num);
-	nfr = posle(num);
-	npost = k(num, nfr);
-	printf("%s\n", dtoa(num, npost, nfr));
-	return 0;
-}
-
-// Запуск программы: CTRL+F5 или меню "Отладка" > "Запуск без отладки"
-// Отладка программы: F5 или меню "Отладка" > "Запустить отладку"
-
-// Советы по началу работы 
-//   1. В окне обозревателя решений можно добавлять файлы и управлять ими.
-//   2. В окне Team Explorer можно подключиться к системе управления версиями.
-//   3. В окне "Выходные данные" можно просматривать выходные данные сборки и другие сообщения.
-//   4. В окне "Список ошибок" можно просматривать ошибки.
-//   5. Последовательно выберите пункты меню "Проект" > "Добавить новый элемент", чтобы создать файлы кода, или "Проект" > "Добавить существующий элемент", чтобы добавить в проект существующие файлы кода.
-//   6. Чтобы снова открыть этот проект позже, выберите пункты меню "Файл" > "Открыть" > "Проект" и выберите SLN-файл.
